[chore] Refactor cerification mail functionality
- Remove expiry time from email token
- Abstract the email verification body, subject and process to a method inside mailer.
-
