+++ helpers/mailer.js
@@ -1,11 +1,18 @@
 import sgMail from '@sendgrid/mail';
 import dotenv from 'dotenv';
+import helper from './helper';
 
 dotenv.config();
 
 sgMail.setApiKey(process.env.SENDGRID_API_KEY);
 
 const mailer = {
+  /**
+   * @description This method is used to send custom emails
+   * @param {String} emailAddress
+   * @param {Object} emailContent
+   * @returns {Boolen} Status of sent email
+   */
   sendCustomMail(emailAddress, emailContent) {
     const { emailSubject, emailBody } = emailContent;
 
@@ -16,7 +23,40 @@ const mailer = {
       html: emailBody
     };
 
-    sgMail.send(msg);
+    // sgMail.send(msg, (error, result) => {
+    //   if (error) {
+    //     // Do something with the error
+    //     return false;
+    //   }
+    //   // Celebrate
+    //   return true;
+    // });
+
+    sgMail
+      .send(msg)
+      .then(() => true)
+      .catch(() => false);
+  },
+  /**
+   * @description This method is used to send a verification mail to a user
+   * @param {String} emailAddress
+   * @param {string} username
+   * @param {string} link
+   * @returns {Boolen} Status of sent email
+   */
+  sendVerificationMail(emailAddress, username, link) {
+    const emailSubject = 'Verify your email on Authors Haven';
+
+    const emailBody = `
+      <div>
+        <h2 style="color: grey">Hello ${username}, Thanks for signing up on heimdal</h2>
+        Please click here to verify your email address, this link expires in two days.
+        <a href="${link}">${link}</a>
+      </div>
+    `;
+
+    const emailContent = { emailSubject, emailBody };
+    mailer.sendCustomMail(emailAddress, emailContent);
   }
 };
