+++ lib/user.js
@@ -8,7 +8,7 @@ const { users } = models;
  */
 class UserModelQuery {
   /**
-    * @param {stirng} email email must be type string
+   * @param {stirng} email email must be type string
    * @return {object} return email if it exist or null if it doesn't.
    */
   static async getUserByEmail(email) {
